Split TabulatedNuOscillator into header and src
This is generally a monolithic library, but it's helpful to expose
the innards for debugging.  The header now exports a (private) API
for debugging.
